fixed travis ci added feature for keeping raw images

+++ src/fabscan/scanner/laserscanner/FSScanProcessor.py
@@ -480,7 +480,9 @@ class FSScanProcessor(FSScanProcessorInterface):
 
         self.eventmanager.broadcast_client_message(FSEvents.ON_INFO_MESSAGE, message)
 
-        self.utils.delete_image_folders(self._prefix)
+        if not bool(self.config.keep_raw_images):
+            self.utils.delete_image_folders(self._prefix)
+
         self.reset_scanner_state()
 
         event = FSEvent()
